requirements.txt
Django==4.2.0
django-tables2==2.6.0
django-filter==23.5
django-crispy-forms==2.0
crispy-bootstrap5==0.7